+++ database/migrations/2024_01_10_131414_add_approval_code_to_package_histories_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('package_histories', function (Blueprint $table) {
-            $table->string('approval_code')->nullable()->after('payment_method_identity');
-        });
+        if (!Schema::hasColumn('package_histories', 'approval_code')) {
+            Schema::table('package_histories', function (Blueprint $table) {
+                $table->string('approval_code')->nullable()->after('payment_method_identity');
+            });
+        }
     }
 
     /**
@@ -21,9 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('package_histories', function (Blueprint $table) {
-            $table->dropColumn('approval_code');
-
-        });
+        if (Schema::hasColumn('package_histories', 'approval_code')) {
+            Schema::table('package_histories', function (Blueprint $table) {
+                $table->dropColumn('approval_code');
+            });
+        }
     }
 };
